# deadCode: use VarEnv to track used binders